requirements.txt
mysql-connector-python
jinja2
python-dotenv
boto3
openai
requests
python-slugify
unidecode
pandas